requirements.txt
django~=3.1.6
Pillow~=8.1.2
psycopg2-binary~=2.8.6
django-bootstrap-v5~=1.0.1
django-jquery~=3.1.0
django-choices~=1.7.1
djangorestframework~=3.12.0
drf-yasg2~=1.19.4
celery~=5.0.0
redis~=3.5.3
django-redis~=4.12.1
gunicorn==20.1.0
whitenoise==5.2.0
python-dotenv~=0.17.0
coverage~=5.5
factory-boy~=3.2.0